uiux: dynamically determine the number of entries to show in each row
the number of entries per row was hard-coded at 3. this change sets the
max width of an entry and uses the max width of the list to determine
how many entries per row. this will also update when the max width
changes.

additionally the rows are determined based on the list row index instead
of building the entire row set and passing that to the list.