makefile: build default way which is to support wayland
i made the decision to build without wayland support because the
decorations were not inline with my other window decorations. i believe
this is due to the gnome devs not implementing server side decorations,
and most other desktops seem to have that.

and to note: building without wayland sacrifices features such as fling
support when scrolling.