refactor: upgrade to gioui.org@v0.5.0
This is an upgrade from v0.3.0, so this also accounts for changes in
v0.4.0 (eg. the new event processing from app.Window, the new Update
methods, etc). The only UX regression I can see so far is that I don't
know how to start the app with the search field focused.